refactor(voice): extract macOS Apple Speech backend to platform_speech/macos.rs
Move the macOS native-status / transcription / FFI layer (native_status,
transcribe_wav_path, availability_from_native, engine_label,
take_c_string, cancel_active_transcription, extern "C" block) into
platform_speech/macos.rs as pub(super); call sites in mod.rs are
qualified as macos:: and cancel_active_transcription is re-exported at
the platform_speech root so voice.rs is unchanged. Cross-platform
contract, shared WAV helpers, stub impl, and tests remain in mod.rs
(411 LOC, down from 971). Behaviour-neutral; check green on darwin.
